requirements_dev.txt
-r ./requirements.txt
-r ./requirements_rtfd.txt

coveralls
cryptography
flake8
jinja2
matplotlib
pyprind
pytest-cov
pytest-flake8
pytest-runner
watchdog
wheel
